requirements.txt
face_recognition
opencv-python
opencv-python-headless
numpy
pillow
pyserial